+++ src/kubernetes/resources/resource.rs
@@ -43,8 +43,15 @@ impl ResourceItem {
     }
 
     /// Creates [`ResourceItem`] from kubernetes [`DynamicObject`].
-    pub fn from(kind: &str, group: &str, crd: Option<&CrdColumns>, stats: &Statistics, object: DynamicObject) -> Self {
-        let data = Some(get_resource_data(kind, group, crd, stats, &object));
+    pub fn from(
+        kind: &str,
+        group: &str,
+        crd: Option<&CrdColumns>,
+        stats: &Statistics,
+        object: DynamicObject,
+        is_filtered: bool,
+    ) -> Self {
+        let data = Some(get_resource_data(kind, group, crd, stats, &object, is_filtered));
         let filter = get_filter_metadata(&object);
         let uid = get_object_uid(&object);
         let creation_timestamp = get_age_time(&object.metadata);
@@ -99,8 +106,8 @@ impl ResourceItem {
     }
 
     /// Returns [`Header`] for provided Kubernetes resource kind.
-    pub fn header(kind: &str, group: &str, crd: Option<&CrdColumns>, has_metrics: bool) -> Header {
-        get_header_data(kind, group, crd, has_metrics)
+    pub fn header(kind: &str, group: &str, crd: Option<&CrdColumns>, has_metrics: bool, is_filtered: bool) -> Header {
+        get_header_data(kind, group, crd, has_metrics, is_filtered)
     }
 
     /// Returns [`TextColors`] for this kubernetes resource considering `theme` and other data.